(PC-23190)[API] chore: extract delete date test
