+++ api/tests/routes/public/individual_offers/v1/delete_date_test.py
@@ -1,0 +1,51 @@
+import pytest
+
+from pcapi.core.offerers import factories as offerers_factories
+from pcapi.core.offers import factories as offers_factories
+
+from . import utils
+
+
+@pytest.mark.usefixtures("db_session")
+class DeleteDateTest:
+    def test_delete_date(self, client):
+        venue, api_key = utils.create_offerer_provider_linked_to_venue()
+        event_offer = offers_factories.EventOfferFactory(
+            venue=venue,
+            lastProvider=api_key.provider,
+        )
+        to_delete_stock = offers_factories.EventStockFactory(offer=event_offer)
+
+        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).delete(
+            f"/public/offers/v1/events/{event_offer.id}/dates/{to_delete_stock.id}",
+        )
+
+        assert response.status_code == 204
+        assert response.json is None
+        assert to_delete_stock.isSoftDeleted is True
+
+    def test_404_if_already_soft_deleted(self, client):
+        venue, api_key = utils.create_offerer_provider_linked_to_venue()
+        event_offer = offers_factories.EventOfferFactory(
+            venue=venue,
+            lastProvider=api_key.provider,
+        )
+        already_deleted_stock = offers_factories.EventStockFactory(offer=event_offer, isSoftDeleted=True)
+
+        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).delete(
+            f"/public/offers/v1/events/{event_offer.id}/dates/{already_deleted_stock.id}",
+        )
+
+        assert response.status_code == 404
+        assert response.json == {"date_id": ["The date could not be found"]}
+
+    def test_404_if_others_offerer_offer(self, client):
+        offerers_factories.ApiKeyFactory()
+        others_stock = offers_factories.EventStockFactory()
+
+        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).delete(
+            f"/public/offers/v1/events/{others_stock.offerId}/dates/{others_stock.id}",
+        )
+
+        assert response.status_code == 404
+        assert response.json == {"event_id": ["The event could not be found"]}
